trace_rca_algo/kb_deploy/requirements.txt
kubernetes
mlflow_skinny
